Update main.go to fix scale to zero functionality
Currently we're checking if the number of connections is greater-or-equal to 1. However the ConnectionCount query that runs doesn't exclude itself from the result set, so it always returns at least one. 

The >= 1 check was added here: https://github.com/fly-apps/postgres-flex/pull/200. This PR reverts back to the previous `>1 ` check.